Allow input generation with resource override + fix defaulted optionals
