+++ janis_core/tool/commandtool.py
@@ -241,7 +241,7 @@ OUTPUTS:
 {outputs}
 """
 
-    def generate_inputs_override(self):
+    def generate_inputs_override(self, with_resource_overrides=False, hints=None):
         """
         Generate the overrides to be used with Janis. Although it may work with
         other
@@ -251,6 +251,18 @@ OUTPUTS:
         for i in self.inputs():
             if not i.input_type.optional or i.default:
                 d[i] = i.default
+
+        if with_resource_overrides:
+            cpus = self.cpus(hints) or 1
+            mem = self.memory(hints)
+            d.update(
+                {
+                    "runtime_memory": mem,
+                    "runtime_cpu": cpus,
+                    # "runtime_disks": None
+                }
+            )
+
         return d
 
     def wrapped_in_wf(self):
