+++ janis_core/translations/cwl.py
@@ -325,60 +325,6 @@ class CwlTranslator(TranslatorBase):
 
         return tool_cwl
 
-    @classmethod
-    def build_resources_input(
-        cls, workflow, hints, max_cores=None, max_mem=None, prefix=None
-    ):
-        from janis_core.workflow.workflow import Workflow
-
-        # returns a list of key, value pairs
-        steps: Dict[str, Optional[Any]] = {}
-
-        if not prefix:
-            prefix = ""
-        else:
-            prefix += "_"
-
-        for s in workflow.step_nodes.values():
-            tool: Tool = s.tool
-
-            if isinstance(tool, CommandTool):
-                tool_pre = prefix + s.id() + "_"
-                cpus = tool.cpus(hints) or 1
-                mem = tool.memory(hints)
-                if max_cores and cpus > max_cores:
-                    Logger.info(
-                        f"Tool '{tool.tool()}' exceeded ({cpus}) max number of cores ({max_cores}), "
-                        "this was dropped to the new maximum"
-                    )
-                    cpus = max_cores
-                if max_mem and mem > max_mem:
-                    Logger.info(
-                        f"Tool '{tool.tool()}' exceeded ({mem} GB) max amount of memory({max_mem} GB), "
-                        "this was dropped to the new maximum"
-                    )
-                    mem = max_mem
-                steps.update(
-                    {
-                        tool_pre + "runtime_memory": mem,
-                        tool_pre + "runtime_cpu": cpus,
-                        # tool_pre + "runtime_disks": None
-                    }
-                )
-            elif isinstance(tool, Workflow):
-                tool_pre = prefix + s.id()
-                steps.update(
-                    cls.build_resources_input(
-                        tool,
-                        hints,
-                        max_cores=max_cores,
-                        max_mem=max_mem,
-                        prefix=tool_pre,
-                    )
-                )
-
-        return steps
-
     @staticmethod
     def workflow_filename(workflow):
         return workflow.id() + ".cwl"
@@ -574,7 +520,7 @@ def translate_step(
     for k in ins:
         inp = ins[k]
         if k not in step.sources:
-            if inp.input_type.optional:
+            if inp.input_type.optional or inp.default:
                 continue
             else:
                 raise Exception(
